libPLE: use same FLAGS as parent Code_Saturne build.
This helps ensure subconfigure is consistent. Bootstrapping also
inherits prior cleaning stage from that of parent.

Also, dependencies on MPI are handled in the standard Automake
manner when building libPLE (this means building a dynamic
libPLE on top of a static MPI may no longer be possible).

This implies changing at least the patch release number for PLE.


git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@3480 e44c5d27-f830-0410-8ce8-974f7a764be6