Various build system fixes and improvements.
Also, dependencies on libraries are now handled in the standard
Automake (this means building a dynamic libsaturne on top of
static dependency may no longer be possible).

Handling of dynamic loading of libraries (for plugins) is reverted
back from rev. 2953, as the standard libtool macro does not
allow building without a dynamic loader if one is found, and
couples dynamic loading to enabling of shared libraries, while
a statically-built executable should be able to load dynamic
libraries.

Detection of dynamic versions of SCOTCH, METIS, and HDF5 is
now made possible.

libbft and libmei are not installed anymore as standalone
libraries (this was a relicate from the separate-library era).

Finally, on Mac OS X, compiling of user subroutines for a
static build requires unarchiving the library and overwriting
selected object files, to avoid issues with multiple
definitions not being handled by the Mac OS X linker.


git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@3481 e44c5d27-f830-0410-8ce8-974f7a764be6