Adapt the Multigrid algorithm to vectorial Poisson equation (such as mesh velocity for ALE).
The aggregation criterium is based on the trace of the diagonal block DA (3x3), but could be a changed (We could test n.DA.n).
The multigrid algorithm for scalars (such as the pressure field) is rigourously unchanged.

Comments in codits/coditv are updated and translated in English. GOTO are replaced by dowhile loop.


git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@3557 e44c5d27-f830-0410-8ce8-974f7a764be6