Improve definition of time moments. - Using the GUI, the moment's name is now based on the user-defined label. - With user functions, allow full access to the API for time moments definition.
Handling of probes is also modified for all fields: a new function in the
entsor.f90 module, field_post_id, returns the index "ipp" of a field in
ihisvr, adding it if not already done. Combined with simplifications
in the GUI data handling, ordering of fields is thus not constrained
by mapping to ipp anymore.


git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@6400 e44c5d27-f830-0410-8ce8-974f7a764be6