Fix Bug in ALE when imposing node displacement.
When imposing the displacement of all the nodes, then the boundary face
was considered as a imposed mesh velocity (ivimpo). Then, the BC of the
fluid was imposed to be the value of the mesh velocity, which is not
wanted when we choose initially that a side is a sliding side (no shear
should be added to the fluid).

git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@7374 e44c5d27-f830-0410-8ce8-974f7a764be6