Combustion: automatic equilibrium of the chemical reaction (provided by B. Sapa)
The user can specify the composition of the reactants and the product composition is calculated
thanks to a atomic balance on C, H, O and N atoms. If other species than CO2, H2O or N2 are present
in the products, the user have to specify the quantity. These species have to be placed after the
Fuel, O2, CO2, H2O and N2.
To use the model, the user set the number of known global species on 2.
If the number of known global species is set at 3, the composition have to be define (previous
model).

git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@8627 e44c5d27-f830-0410-8ce8-974f7a764be6