Fix for OpenMP multigrid-preconditioned Conjugate Gradient convergence issue.
When running with multiple threads, race conditions in the Gauss-Seidel
smoother or not of themselves in issue (i.e. the algorithm simply degrades
to Jacobi at rank/thread boundaries), but do amount to preconditioning
with a slightly different matrix at each iteration, which requires a flexible
PCG variant to avoid occasional convergence issues.