Work around missing debug symbols issue with some gcc versions and builds.
This occurs when gcc only links libraries, but no .o files.
This workaround will fail when using "code_saturne compile --force" when
no user files are present, but as this is not a standard case, it
can be mostly ignored.

A better way of handling all situations would be to compile and install
cs_solver.o in the "lib" directory, but automake and libtool are not
helpful for this, so this may require more work on the Python side.