Fix handling of ALE mesh viscosity setting in restart configurations.
The mesh viscosity is not saved in the checkpoint file anymore, as it
can normally be rebuilt. For really advanced cases where it would need to
be saved, the user could force the behavior anyways as this is a standard
property field. The user documentation has been partially updated
(but a move to Doxygen for this part would be better).