Fix c17333d (2018-06-22) in case of restart with auxiliary restart file. The read total pressure was used to reinitialize the solved pressure without reading the reference point, leading to an erroneous solved pressure.
The reference point is now read in auxiliary if it had been located by the code
or specified by the user in the previous calculation.
In case of restart with auxiliary, solved pressure and total pressure are read
and not recomputed.

The reference point is considered non initialized if at least one of its
coordinate is non-initialized.