SALOME_CFD: fix (or so it seems) dissapearing windows bug on current systems.
Issue seems solved on Debian 9 with Salome 9.2.

The solved issue behaves as follows: whenever the focus of a Code_Saturne
GUI page is modified (either switching windows with external tools or
using a function calling a popup, such as the MEI editor), the page would
be reduced and difficult to bring back to its position if it contains
a QTreeview... (observed on Debian 9 and Ubuntu 16.04 at least).