Major fix for open-boundary boundary conditions. Mainly impact the new atmospheric automatic treatment for open-boundaries.
A temporary BC code is creates for the pressure: icodcl = -1
to specify that the pressure is imposed in solved pressure directly (not
in total pressure) and icodcl = 1 is set automatically after.

This might impact other BCs: a first pass before the time stepping in
called so that BC types (itypfb) and reference point for pressure is
computed.

Cherry-picked from ca561696, 478ec368, 220cd690, 3eb58165 and cb503c5c