Allow --with-catalyst=salome configure option when Salome includes Catalyst.
Cherry-picked from commit 7ce856e9 on master.

This is only possible if Salome defines the CATALYST_ROOT_DIR path,
which is currently the case with Salome_CFD. CATALYST_ROOT_DIR must point
to an installation directory of a version of ParaView or Catalyst build
with development headings, MPI support, and Python support.

This solves GitHub issue #29.