Build: use correct option for library relase numbering.
The -version-info option used before corresponds to a different
numbering logic, which is more stringent, requires more frequent
updates, and did not lead to the desired numbering logic.

To avoid causing unnecessary incompatibilities, we should also switch
to this logic for libple, but only upon the next major or minor
revision change.