Work around yet another Intel compiler issue.
Cherry-picked from commit bbe33e0ab on master.

With icc 17.0.0 20160721, we have a crash in __intel_avx_rep_memcpy in
the particle tracking, where adding the braces to the calling loop seems
to avoid the issue.

Strange way to fix a bug, and stranger yet that icc does not seem
to parse the file correctly...