[CDO] Scalar solve: Choice on RHS reduction
By default, a cs_interface_set_sum() on the RHS was performed before
solving a linear system. That's why all functions have true as
parameter.
In order to prepare future integrations, a parameter is added to have
the choice to perform or not this operation.
This may have an impact on parallel computations.