Remove some newly added extended neighborhood options which had limited impact.
On a given (difficult) test case, only the full extended neighborhood and
"cell center opposite" options avoided a crash so the new additions with lower
impact are not considered to have sufficient value. The legacy extneded
neighborhood also fails, but is kept to allow for setups equivalent to those
of older version, though it should be considered deprecated.