Do not output boundary temperature in case of minimal XML file and enthalpy.
This avoids issues when the boundary enthelpy to temperature file is not
defined. A better solution would be to base the default on the availability
of the enthalpy to temperature conversion, but this is not easy to check
(though it would be easy to check using a function pointer mechanism).