Fixes for GUI-defined Lagrangian statistics output.
Cherry-picked from commit d9b06f00 on master.

This should fix the root cause of GitLab issue #152.

Remames were missing, leading to a crash in the code when trying to use
the associated fields for time moments (whether it is a good idea or not
to build time moments in these statistics may depend on whether they
are defined as unsteady). The fix involved handling renames in the
XML update, but also updating the available fields list.

To avoid changing default behavior relative to an "empty" XML file,
as statistics may be known by default for a given model but deactivated,
a "status" attribute was needed to allow for deactivated field definitions,
and handling this in the output control. This also required minor additions
to the XML engine code, so as leave the attribute optional.