Finalize switch to cs_all_to_all API in cs_block_to_part.
Using the existing cs_all_to_all API, this requires an extra exchange compared
to the previous API in cs_block_to_part_create_by_adj_s, whereas this could be
optimized by adding an optional indirection array when building the
distribution structure, or add a function to direclty create a distributor
in reverse mode. Still, the total number of exchanges is usually only increased
by a few percent, and possibilities of optimizing the exchanges inside the
API should outweigh this.