[Atmo] add 3D radiative transfer for solar wave (shortwaves): direct and diffuse.
The 1D atmospheric model is solved before to get boundary conditions at
the top of the CFD domain. It is also used for the absorption
coefficients.

This work is part of M. Millez PhD, Y. Qu PhD and Y. Maanane internship.

Also rename verbose mode for Radiative transfer (iimlum -> verbosity).