Allow modifying boundary faces so as to ignore some of them in the computation.
This should be useful mostly to speed-up 1d or 2d models where most
boundary faces are symmetry faces and have no contribution to the
Finite Volume scheme.

In this case, we must still take care to remove the selected faces from
the boundary mesh postprocessing output (replacing the "all[]" default
selection criteria with "not symmetry" for example). Doing this
automatically would require being able to rebuid volume cells in another
manner, or managing faces to ignore in another manner (for example
as a special symmetry zone, handled in a more integrated manner).

Currently, only the Green-Gauss based gradient reconstruction
with iterative handling o non-orthogonalities may be used with this,
as the least squares gradients may fail due to insufficient
covariance directions. This could be improved by automatically
adding "cocg" entries in the normal directions when all contributions
are on a same plane or axis, or using the "hidden" symmetry boundary
cells in this step. The first solution woud be preferred as adaptable to
true 1d or 2d meshes.