code_aster coupling: updates for salome_meca 2019.
There is still an error when running with salome_meca that seems to be related
to a change inside the CALC_IFS_DNL operator in code_aster.
The code runs with salome_meca 2017.0.2 (on recent systems and a build of
salome_meca based on older compilers, it may be necessary to add:

SALOME_KERNEL_CXXFLAGS="-std=c++03 -D_GLIBCXX_USE_CXX11_ABI=0"

to the configure command.