Move some elements of the user documentation to Doxygen.
This is a first step, which can be used as a starting point and example,
but does not imply a "final" organization.

The case hierarchy described anticipates the planned work on
the "run.cfg" replacement to "runcase" files.

Note that .h files could be renamed to .dox or .txt to better identify them,
and for simplicity and better reuse options, Markdown can be used.
Both are used here (mainpage.h renamed to mainpage.dox, and case_structure.md
added).