Bump minimal required Python version to 3.4.
Python-2 compatibilty is not actually removed at this stage.
Only the configuration testing requirements and documentation
require Python 3.4, so it is possible to revert this decision
in case of negative feedback.

Still, Python 3.4 was released in 2014, and is packaged on older
systems still somewhat common such as RHEL or CentOS 6.0.
Anything older is really obsolete...