[CDO] (MAJOR) Reshape the management of high-level iterative algorithms
s.t. Picard, Uzawa or Golub-Kahan Bi-orthogonalization
This commit modifies advanced user settings for Stokes/Navier-Stokes
equations

Several new keys have been added to get a finer control of the
tolerances used when solving either the non-linear process or the
inner linear system.
Same thing for the verbosity.

The consequence is a modification of the way of settings some
tolerance coefficients for outer or inner algorithms in
cs_user_parameters()

This commit prepares the add of new algorithms to solve the
non-linearity process.