Add experimental --enable-long-lnum configure option, with major edits.
This requires major cleanup in usage of in/cs_int_t/cs_lnum_t arguments.
At this stage, the code compiles without additional warnings, and has been
lightly tested on "mesh verification/qauliy mode". It is not compatible
with the Fortran API, as we do not use different integer kinds yet
(this would be feasible, but moving to C is the preferred solution).