[Studymanager GUI] Avoid crash when a case has multiple postprocessing inputs.
To handle multiple inputs correctly, it would be necessary to modify the
presentation so as to handle a list or table, with addition and removal.

Here, as a partial fix to GitLab issue #107, when multiple input nodes
are present in a case, "<multiple inputs>" is displaye, and the matching
selection button is disabled. This at leas avoids a crash of the GUI.