Avoid update issue when restarting with an ALE mesh.
Cherry-picked from commits d44d895a and 33cae87b on master.

A cleaner solution is to also save the previous time step, but as
this has forward compatibility issues, it will be done separately to
as to allow merging this patch on older versions.

Cleaner handling or restart for ALE computations.