Convection-diffusion multigrid: restore original aggregation criteria.
This original criteria was modified upon integration of this solver
as it seemed to strangely limit the grid hierarchy depth on a simple
test-case. It seems to provide much better performance on Rij-epsilon-EBRSM
test cases during the initial iterations (with no purely diffusive
zone, but we assume a solver adapted to both convection and diffusion
should behave well in convective zones).