Major commit in hydrostatic pressure computation for improved hydrostatic algorithm (iphdr=1).
Hydrostatic pressure is now a field, so convergence is in the
run_solver.log.
It is not solved in increment, so the precision required is more
coherent with the one required for the pressure. And time iteration will
help reconstruction loops.
This patch should save time in the atmospheric module (where iphydr=1 is
the defautl algorithm and this step is really time consuming).