Move cs_gui_user_boundary_conditions_define before cs_user_finalize_setup.
This is useful so that if fields (such as roughness) are defined based
on the boundary conditions, they will be available in cs_user_finalize_setup
in case the user wants to modify their logging and post-processing
attributes.

This will also be required when we add "xdef" type BC definitions, so
that the definitions provided by the GUI are present before those defined
by the user in cs_user_finalize_setup.