[Atmo] Fix bug introduced in 69e7849 (2020-08-25).
Atmo initialization of variables should be done in two stages, before
and after the GUI.
The variable "initmeteo" should be ideally set by the GUI, to know if
variables are set from a meteo profile or from the GUI formula.
Here we chose to use the meteo profile if the user specify one, use the
GUI otherwise.