Add AMGX-based linear equation solver.
This currently requires that matrices be in CSR or MSR format, and in parallel,
that cs_renumber_set_algorithm() be called with halo-adjacent cells last.

The code has not been fully debugged/tested yet, but is ready for that stage,
pending available hardware and CUDA environments.

At this stage, some issues related to the port remain:

- When using the solver configuration from the default string, the
  AMGX solver configuration does not seem usable, at least not when running
  on the host.
- Most solver configurations from file tested lead to a "non-implmented"
  error when runing on host, though they are available on the device.
- Passing the diagonal matrix elements to AMGX_matrix_upload_all leads
  to an error, so although MSR matrices are allowed in theory, they
  are not usable yet.
- On a simple test case, AMGX_vector_upload fails, but this might be
  due to a compatibility issue (tested on CUDA 11, while the standard
  AMGX sources support no higher than CUDA10), so should be tested
  on another machine.

Other issues are related to the API's logic:

- The solver tolerance must be set in the solver configuration, and
  the value passed through the cs_sles_solve function is ignored.
- long cs_lnum_t types are not yet supported in these bindings.
- The AMGX API only supports integer type indexes, so matrices with
  more than 2^32 rows are not handled.

Periodicity is not handled yet either, at least not in matrix halos,
so periodicity will only be handled for matrices built through
the cs_matrix_assembler API.

Parells has not been tested yet, due to remaining issues.