Further improvements to AmgX solver.
In case of periodicity of translation with AmgX, for "legacy" finite-volumes,
the matrix is now assembled using cs_matrix_assembler so as to
avoid the periodicity-specific halo sections.

This could be used for other external libraries such as PETc, but is will
be preferable to reorganize their assembly so as to avoid another copy.