[GUI] Add button and pop-up window for the thermodynamic pressure.
We may move this button in the future.

Manage this in the GUI and the xml update (smart behavior if the user
changes main module, retuens to combustion etc.).

Remark: need to do some mofications in order to remove from the xml
"thermodynamical_pressure" when the combustion module is not selected
(it is only set to "off" curently).