[GUI] Add gas combustion variables in the panel 'Species transport".
The variance must be separated and the minus button must not create errors.
To do so, I needed to create a new type for variable model (var_model type) was
needed. Also, I needed to change the part for the clip of mixture_fraction and
the variance.

In fact the original "model" type was used to skip the model lines in the xml
to avoid the clip (see getClippingList()). So, the type is no more 3dp
(or other) for the variable parameters (before our changes, it was 'model'),
now it becomes 'var_model' to give the possibility to put a clip on
these variable added by the combustion model.