Added mesh modification flag types to more easily recommend re-partitioning.
The existing system was based mainly on the face-cells graph update
due to mesh joining, but this was not sufficient anymore.
The new system can be refined, and repartitionning is not recommended
automatically, the user is assumed to have better knowledge of whether
many elements are impacted or not. This could be automated with heuristics
or additional flags in the future.

The existing hint system for mesh joining may also be improved to use
the same mechanism.