(Major) Make more flexible the usage of cs_param_sles_t structure
This is not anymore associated to a field id or an equation but can be
used to store the settings of auxiliary linear algebra systems.
This can be used to define complex algorithms to solve saddle-point
systems with a Schur complement approximation or block preconditioners.

Some advanced user settings related to the linear algebra can be
impacted.