Lagrangian module: keep rotation matrix for reverse change of reference frame.
This simplifies the handling of multiple cases. It seems a bug was present in
commit e53f1372 relative to this, as the rotation matrix for the general "sphere"
model is transposed twice: one when computing its coefficients, a second time
when using the `cs_math_33t_3_product` function instead of `cs_math_33_3_product`
as in the direct transformation.

This is assumed to be an error. The rotation is probably still incorrect, as
the matrix expressions for the new CS_LAGR_SHAPE_SPHERE_MODEL are the
reverse/transpose of those used for the CS_LAGR_SHAPE_SPHEROID_STOC_MODEL,
which was purportedly fixed in commit 1795093a.