Scripts: try to remove additional warnings due to lmod.
These warnings are due to some multiline BASH_FUNC_m* environment variables,
which are not handled or incomplete in the sub-environment, so these
variables are removed.

When a bash shell is used, these variables are temporarily restored when
calling a bash script or a run prologue or epilogue, but could probably
be simply removed, as it is not clear whether they are of potential use
or not in these sections.