Revert GUI automation part of commit e9e23b72.
Automatic changes in the GUI with no conversion path caused issues;
We need to add temperature Dirichlet as an additional (not replacement)
option, so as not to break some case setups, and allow a more progressive
migration. There are several options at the XML level:
- add BC's on temperature, which is a property, not a variable, when
  the thermal modle is Enthalpy
- add a dirichlet_temperature type
- add an additional 'convert_from' optional flag to boundary conditions
The last option is probably the simplest to implement an generalize.