Add gather/scatter functions to exchange a global vector from all ranks to a root rank
Rely on the adaptation of cs_parall_allgather_r function
Useful to handle the right-hand side within the MUMPS interface.

There are a double-precision and a single-precision variant along with
a gather or scatter operation.
These functions imply that a global array is stored on a root rank,
thus limitating scalabity.

Use with caution.