Probes: improve snap to cell mode robustness.
Only one probe per cell is kept, even with probes along a cell boundary
alternatingly one one side or the other. This behavior might be too aggressive
with complex profile shapes passing multiple times in a cell, but snap
mode is probably not the best choice for those exceedingly rare
(if not purely conceptual) cases anyways.