Activate fluid/solid flag independently of internal coupling.
Using the GUI, as soon as solid zones are defined, internal
coupling by separation of zones is still activated. This could
be improved in the future.

At lower levels, the two functionalities are less intertwined.

TODO: in the GUI, we should check if solid zones are present
in the internal coupling view, rather than manage a list of internal
coupling zones from the zone type view. This is simpler.
Presently, if we add a zone in the Mesh section, define it as solid, then
rm remove it in the Mesh section, this is not propagated to
the internal coupling model. This is a bug and should be fixed.