[compressible] Fix commit 3a4e215 (September 15, 2017).
Cherry picked from commits 4252e31e, 9eed94f9, and 7d4ee7fb on master.

When using the imposed inlet/outlet boundary condition, the boundary mass flux
was not computed using the Rusanov scheme (i.e. was not consistant with
momentum and energy convective fluxes on the boundary with this condition).