boundary_thermal_flux function: use C balance function.
Cherry picked from commit 71dc8ee1 on master.

In case of internal coupling, the previous computation seemed to be off
by a constant or external reference, with Q1 on the solid side, Q2 on the
fluid side, and Q1 != Q2. But Q1-Q2 matching the result of the C balance.

TODO: apply the same logic to compute the boundary Nusselt.