Compressible model: precompute boundary density for imposed mass flow BC.
This allows prescribing mass flow, pressure, and temperature for inlets.

Mass flow still cannot be prescribed when density is computed using
the total energy, as this requires knowledge of the velocity, but the velocity
needed for a given mass flow depends on the density, so an iterative
process would be needed in that case.