Major commit: migrate enthalpy/temperature conversion to C, with user control.
Cherry picked from commits 8b7b1185, d20e235b, and b51f246c on master.

Some physical models are still based on Fortran code, but the base API
is now in C, and user-defined functions are zone-based.

Since some boundary temperature to enthalpy computations are still called
with an arbitrary (not zone-based) face ids list, extra computation, and
an extra copy for filtering are required, but the API is prepared for migration
to zone-based boundary conditions.