Fix use of uninitialized verbosity.
With commit 961e7ed0, the equation parameters for velocity were moved,
and not available for tests on the associated verbosity.
But verbosity of velocity should not be used for other variables anyways,
so the test is simplified.

The logging of properties here could probably be removed, as it is handled
by the general field logging mechanism.