Initialize variable_id, scalar_id and reference_diffusivity for user fields.
Cherry picked from commit df647a5e on master.

Prior to this fix, user variables defined in C or the GUI were missing
the variable_id and scalar_id key values (as well as iscal and ivarfl entries),
and the reference diffusivity was not always located in the XML file.