Improvements to cs_malloc_hd API.
A `cs_set_alloc_mode` function was added so as to allow modifying the
allocation type on-demand, and a general CS_FREE macro and bft_mem API
alternative function pointer mechanism is set in place so as to allow
mixing the base bft_mem_malloc and cs_malloc_hd APIs.

Not mixing the APIs is still recommended, but the new features should make
it much easier to use arrays built in other portions of the code on an
accelerated device without requiring deep changes in most of the code.