Scripts: cs_package now loads configuration from file.
This is based on the information provided by the
`lib/code_saturne_build/cfg` file.

This should help avoid reload issues when searching for alternate
packages, provides a single stop for build information, and should
make it easier for post-installation scripts to modify the configuration
if really required, without patching Python files.

This could also allow external tools not based on Python to query
the configuration information.