Remove support os symmetric-structured CSR matrices.
This was never used in practice, and difficult to parallellize
(in the shared-memory or GPU sense), so did not provide any
performance gain.

It was kept for a time in case external libraries could benefit from
this, but most known libraries either do not support symmetric CSR
matrices (for example HYPRE), or do not recommend them for performance
reasons (for example cuSPARSE).