HYPRE: now really works on GPU.
With the latest HYPRE version (2.23), this requires allocating the
matrix and vector values passed to HYPRE on the GPU (at least
using shared/managed memory). This copy could maybe be taken advantage of
in the future for matrix assembly using an asynchronous process.
For vectors, this limitation of HYPRE is added overhead, though some
commments in HYPRE's source code seem to indicate the situation
may be temporary and fixed in future versions.