HYPRE: add support/translation for legacy block matrices.
These are handled as regular scalar matrices on the HYPRE side.
For matrices with block diagonals, only nonzero block elements
are added, as blocks for non-boundary cells usually contain only
zeroes off-diagonal.

For full block matrices, it is not clear whether HYPRE supports those
directly or only internally (they are not mentioned in the documentation)
so we also convert the matrices to sclar matrices.