Progress on accelerator support: ghost cell exchange on device buffers.
For accelerator device, the objective is to gather the send buffer
directly on the device. With CUDA-aware MPI or similar, the exchanges
can then be called directly on the device. Otherwise, unless we use shared
memory, we use an additional "receive buffer", whose data is the copied
to the device. This allows copying only the halo values and not the full
array. For shared memory, prefetch is called only for the ghost values.

The code is tested for non-regression on the host, not tested on
accelerators yet.