[Solidification] Enforce pressure and face velocity for solid cells.
Case of MUMPS applied to the full system. When an iterative algorithm
is used (for instance GCR or ALU) there is no need to prescribe the
pressure inside solid cells (the value remains to its initial value).
Moreover, one needs to penalize the face velocity to ensure a good
incompressibility enforcement of the velocity field.