GUI reader: now handle "neumann_formula" using xdef and cs_dof_func_t.
A bit of addtional cleanup is done also, and consistency is improved
so when a MEG-defined value is already associated to a variable and
boundary condition, the GUI does not overwrite it (as completing
default by the code should be done at a later stage, and we assume
existing settings are forced by the user).

This is done for scalars and the inlet velocity, but should be done
in all cases in the future.