Ensure dttens field is always created during setup phase if needed.
Tests are now based on presence of zones requiring head losses, not the
actual number of cells.

This also allows adding a check in fldini when defining field gradient
weights, to ensure we do not have conflicting definitions from different
models.

This also avoids cases where dttens was present but not adressable in user
defined functions of the setup stage, and did not appear in the setup.log.