Multigrid: make mesh association for smoothing more generic.
The association is now assigned to the matrix (optionally),
and could represent some form of dual elements rather than
only primary cell and face quantities.

When face normals are not available for a given level, use
cell center distances instead for smoothing (this may happen
for example using an assembler on the primary mesh, with
internal coupling).

Also pass additional adjacency info so as to prepare switch
from native to MSR as the primary matrix form (and use "native"
structures only inside the multigrid algorithms, and possible
migrate those steps to MSR (avoiding an extra copy) at a later
stage.