CUDA: some fixes.
For SpMV producs, we force device synchronization, and assume x and y
arrays are provided on the host.

In the future, it would be interesting to separate the synchronization
(adding a wrapper in cs_base_accel.c for cudaDeviceSynchronize or
its oneAPI or OpenMP equivalent, if present), and to provide a
cs_matrix_get_alloc_mode  query function so as to align the behavior
of x and y. This would make things a bit more verbose and cumbersome
in the benchmark mode code, but would reduce synchronization in the
actual code whre SpMV's are used.