Fix matrix assembly with periodicity following commit fcdcc07a (March 29 2022).
This fix ensures the global ids used are "stable" and not changed by use
of a different range set, but may need to be improved, so that in case
of periodicity, the range set is built earlier and used consistently.
The current fix at least allows use of AmgX with periodic meshes.