[GUI] Complete commit 6915e608 with separate batch template and documentation.
For a given batch.<TYPE> template, the associated documentation will be in
the associated batch_help.<TYPE> file.

This allows providing detailed help in the GUI (with the added advantage
of using HTML if desired, as is done now for SLURM), while keeping
batch templates added in the run.cfg file clean.