Various fixes for refinement and coarsening.
Face refinement generation (cell refinement level at which and interior face
is added) and refinement level (number of subdivisions relative to root parent)
should not be confused.

This also means that contrary to the message of commit 6e44e85, rebuilding
face_r_gen based on vtx_r_gen should be removed from the planned steps.

Using the face refinment levels computed by the local
_compute_face_r_level function in cs_mesh_refine.c, it should be possible to
use this information (or apply a similar logic) to locally rebuild the
parent quadrangles and triangles of cells with a mix of already refined and
non-refined faces, so as to obtain cleaner refinement patterns.