[GUI] Add fine-grained activation of initialization formulas.
Based on code by M. Lab, completed by Y. Fournier, with additional
cleanup and refactoring (including of the older code base).

This should solve GitLAb issue #251.

This may be further improved if we add a better (automatic) velocity
initialization option, in which case we may then replace the check box
with a combo box so as to handel more cases.