Modify name-to-id map so that pointers to allocated names remain stable.
Cherry picked from commit f4872d09 on master.

Previously, increasing the size of the map could lead to reallocation,
and extenal pointers to names in a map could need to be shifted.
This switces to a behavior where names are simply added to buffer blocks
of increasing size, which are never reallocated during the lifetime of the
cs_map object.

The new allocation may be a little less cache friendly than the previous
behavior. On the `tests/cs_map_test` unit test timings, we remain under
the 1e-7 second range for a query, but seemingly more often in the
5e-8 et 8e-8 second range compared to 1.2e-8 to 2e-8 which were more
frequent before (though measurements are both noisy).
In any case, this should not have a significant overall performance
impact, and is much safer.