Remove Green-Gauss with vertex interpolated face values gradient algorithm.
This algorithm was introducted in v7.0 as a faster alternative to the
least-squares algorithm with full extended neighborhood, but was not usable
(i.e. did not provide correct behavior) for the "iphydr" (external forces)
option, where is was most needed.

Promising improvements to the reduced extended neighborhood selection
provide a better and faster alternative (see GitLAb issue #358).