Automatically decide whether LSQ gradient boundary COCG should be recomputed.
In general, the only case where we can use the previously computed COCG
values (dependent on boundary conditions) without recomputing are when we solve
by increments. This is done through the `cs_gradient_scalar_synced_input`
functions, with inc = 0. In this case, we check whether the last solved variable
is the same or not to see if we can avoid recomputing COCG.

Note the recomputing COCG when not necessary only induces a slight performance
penalty, while not updating it where necessary can lead to incorrect gradient
values, so we prefer to err on the side of caution.