Add utility functions for gradients and reconstruction at boundaries.
The functions added here allow reconstructing values at the I' points
associated to boundary faces, for a given selection of boundary faces.

As reconstruction from I to I' involves mostly the transverse component
of the gradient, a Neumann condition can be assumed if the boundary condition
is not well known. This will allow a 2-step approach for faces with
internal coupling, where values can first be reconstructed at I'
(using the functions added here), then equivalent boundary conditions
can be defined at the coupled boundary face centers (to be done).
.