[Atmo] Major commit in the soil model.
It is partially moved to C using field structure and zone structure.
The Boundary conditions on temperature and humidity is not done
automatically.
The use of fields will simplify post-processing and standard user
setting.

To activate it:
  cs_glob_atmo_option->soil_model = 1; /* Switch on soil model */

The user have to chose the category among:
  - CS_ATMO_SOIL_5_CAT
  - CS_ATMO_SOIL_7_CAT
in cs_glob_atmo_option->soil_cat, and chose which zone is concerned, e.g.:
  cs_glob_atmo_option->soil_zone_id = cs_boundary_zone_by_name("Sol")->id;