Remove uneeded initialization to fix compilation, broken by e00612c7 (today).
We could add a dependency to <string.h>, but as cs_math.h is low-level
code_we prefer not to.

We could also simply use a double loop or a cast and keep the
initialization, but as this function is inlined, we prefer to keep it
as short as possible.

Finally, this function is not so short, so not inlining it
(or not forcing the inline static) might be an option.