[GUI] Create NeptuneFieldModel object
This a major modification and still a work-in-progress. The objective is
to reduce the size and complexity of the MainFieldsModel object. This
commit is a first step in this direction as it isolates methods related to
a single field.

Main changes:
* Field specific methods from MainFieldsModel are reimplemented in
NeptuneFieldModel (and should eventually replace them).
* Field index is now independent from the row index in the tableView of
the "Main Fields" page. This means that field addition or deletion does
not modify existing field indexes.

TODO:
* Replace wherever possible the calls to MainFieldsModel by calls to
NeptuneFieldModel.
* Some field attributes interact in a complex way with the other fields.
These interactions are not easy to include in the NeptuneFieldModel
object.