Fix handling of `group_name` option for `cs_mesh_remove_cells` function.
Previously, the group was only added if no non-default was already present,
while the previous boundary group was propagated first when only one
cell layer was removed near a boundary.

It could be useful to allow combining both approaches, but this would require
another option argumnent to that function. In addition, previous boundary
group propagation was limited in the sense that for cells with multiple
boundary faces, the groups associated to an arbitrary face ware propagated
(based on the highest face family number). So we now assume that if users
need adavanced group propagation, they can handle it using additional
processing loops, and the simplest behavioir (priority to the group name if
providded is now the default.

This fixes issue  #410.