Update README and installer relative to ParMetis availability.
ParMETIS has moved to GitHub, but there is no stable release yet, and the old
releases may only be found on third-party sites.

In addition its licence is still quite restrictive, so installation is not
proposed by default anymore.

As this was not the default partitioner anyways, most users will not notice
any practical difference.