Fix crash and stopping criterai with control file and variable time step.
Defining `max_time_value` in a `control_file` with an adative time step
(`idtvar=1`) could cause an exit with no error message in `ddtvar`, due
to a test for coupling which did not seem necessary.

Another issue followed, where the code stpeed cleanly, bit to early, due to
tests on the time step number (and not value) which needed to be handled.