code_aster coupling: remove filtering of force directions.
This was limited to x, y, z directions and could thus not handle rotated
meshes, and if we want to avoid displacement in some directions,
that should be done through the code_aster boundary condition definitions.

On the GUI side, this simplification removes the need for the coupled
structures table, but we keep the (now empty) external coupling tab,
as it can be later used to add global settings, such as verbosity and
visualization levels.